pull out fish component and interfaces
